requirements.txt
pandas==2.2.3
pymongo==4.10.1
tqdm==4.66.6
dask==2024.10.0
python-dotenv==1.0.1